pollination: conform to 5-file spec + rewire onto crop_benefits chain (WIP)
Add pollination_initialize.add_pollination_tasks(p, parent=None), run_pollination.py,
and pollination_method.qmd. Rewrite pollination_tasks/functions to run William
Sidemo-Holm's crop_benefits chain on our SEALS 300m maps at each SEALS year
(configure_crop_benefits off base_data/crop_benefits/). Archive the old 65-crop
script set under _archive_pre_crop_benefits/.

WIP: not yet smoke-tested or attached to a consumer.
